添加国际化信息

+++ centit-dde-core/src/main/java/com/centit/dde/dataset/FileDataSet.java
@@ -61,7 +61,7 @@ public class FileDataSet extends DataSet {
         }
     }
 
-    public void setFileContent(String fileName, int fileSize, Object fileData){
+    public void setFileContent(String fileName, long fileSize, Object fileData){
         if(this.data == null || !(this.data instanceof Map)){
             this.data = new HashMap<>(20);
         }
